refactor: change sequence typing data from int to long to work with very large wMLST and cgMLST data

+++ src/main/java/optimalarborescence/sequences/SequenceTypingData.java
@@ -4,13 +4,13 @@ import java.util.List;
 
 /**
  * SequenceTypingData represents a sequence of typing data, typically used in bioinformatics.
- * It extends the generic Sequence class with an integer array as its data type.
+ * It extends the generic Sequence class with a long array as its data type.
  */
-public class SequenceTypingData extends Sequence<Integer> {
+public class SequenceTypingData extends Sequence<Long> {
 
-    private static final List<Integer> MISSING_DATA_SYMBOLS = List.of(-1, 0);
+    private static final List<Long> MISSING_DATA_SYMBOLS = List.of(-1L, 0L);
 
-    public SequenceTypingData(Integer[] data, int length) {
+    public SequenceTypingData(Long[] data, int length) {
         super(data, length);
     }
 
@@ -18,7 +18,7 @@ public class SequenceTypingData extends Sequence<Integer> {
      * Gets the element at the specified index in the sequence.
      */
     @Override
-    public Integer getElementAt(int index) {
+    public Long getElementAt(int index) {
         if (index < 0 || index >= getLength()) {
             throw new IndexOutOfBoundsException("Index: " + index + ", Size: " + getLength());
         }
@@ -32,8 +32,8 @@ public class SequenceTypingData extends Sequence<Integer> {
             throw new IllegalArgumentException("Cannot compare SequenceTypingData with different Sequence type");
         }
 
-        Integer thisElement = getElementAt(index);
-        Integer otherElement = (Integer) other.getElementAt(index);
+        Long thisElement = getElementAt(index);
+        Long otherElement = (Long) other.getElementAt(index);
         return thisElement.compareTo(otherElement);
     }
 
@@ -46,7 +46,7 @@ public class SequenceTypingData extends Sequence<Integer> {
     public String toString() {
         StringBuilder sb = new StringBuilder();
         sb.append("SequenceTypingData{data=[");;
-        Integer[] data = getData();
+        Long[] data = getData();
         for (int i = 0; i < data.length; i++) {
             sb.append(data[i]);
             if (i < data.length - 1) {
@@ -65,8 +65,8 @@ public class SequenceTypingData extends Sequence<Integer> {
         SequenceTypingData other = (SequenceTypingData) obj;
         if (getLength() != other.getLength()) return false;
 
-        Integer[] data1 = getData();
-        Integer[] data2 = other.getData();
+        Long[] data1 = getData();
+        Long[] data2 = other.getData();
         for (int i = 0; i < getLength(); i++) {
             if (!data1[i].equals(data2[i])) {
                 return false;
